+++ packages/Input/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sipe-team/input
 
+## 0.0.3
+
+### Patch Changes
+
+- f425309: fix: exports
+- f425309: fix: styles.css
+
+## 0.0.3-next.1
+
+### Patch Changes
+
+- fix: exports
+
+## 0.0.3-next.0
+
+### Patch Changes
+
+- 27c312a: fix: styles.css
+
 ## 0.0.2
 
 ### Patch Changes