+++ packages/badge/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sipe-team/badge
 
+## 0.0.3
+
+### Patch Changes
+
+- f425309: fix: styles.css
+- Updated dependencies [f425309]
+- Updated dependencies [f425309]
+  - @sipe-team/typography@0.0.4
+
+## 0.0.3-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @sipe-team/typography@0.0.4-next.1
+
+## 0.0.3-next.0
+
+### Patch Changes
+
+- 27c312a: fix: styles.css
+- Updated dependencies [27c312a]
+  - @sipe-team/typography@0.0.4-next.0
+
 ## 0.0.2
 
 ### Patch Changes
@@ -7,7 +31,6 @@
 - e6f76c0: fix: css module bundle
 - Updated dependencies [e6f76c0]
   - @sipe-team/typography@0.0.3
-
 
 ## 0.0.2-next.0