+++ packages/card/CHANGELOG.md
@@ -1,9 +1,30 @@
 # @sipe-team/card
 
+## 0.0.2
+
+### Patch Changes
+
+- f425309: fix: exports
+- f425309: fix: styles.css
+  - @sipe-team/tokens@0.1.0
+
+## 0.0.2-next.1
+
+### Patch Changes
+
+- fix: exports
+  - @sipe-team/tokens@0.1.0
+
+## 0.0.2-next.0
+
+### Patch Changes
+
+- 27c312a: fix: styles.css
+  - @sipe-team/tokens@0.1.0
+
 ## 0.0.1
 
 ### Patch Changes
 
 - e6f76c0: fix: css module bundle
   - @sipe-team/tokens@0.1.0
-